did refactoring with new abstract datasouse & component
