+++ src/app/local-events/local-events-datasource.ts
@@ -1,4 +1,4 @@
-import { BaseEventsDataSource } from '../base-events-datasource';
+import { AbstractEventsDataSource } from '../abstract-events-datasource';
 import { EventInfo } from '../models';
 import EVENTS from '../../assets/model/local-events.json';
 
@@ -9,11 +9,25 @@ export { EventInfo };
  * encapsulate all logic for fetching and manipulating the displayed data
  * (including sorting, pagination, and filtering).
  */
-export class LocalEventsDataSource extends BaseEventsDataSource<EventInfo> {
-
-  data: EventInfo[] = EVENTS;
-
+export class LocalEventsDataSource extends AbstractEventsDataSource<EventInfo> {
   constructor() {
-    super();
+    super(eventData());
   }
+}
+
+function eventData(): EventInfo[] {
+  const events: EventInfo[] = [];
+  const now = Date.now();
+  for (const event of EVENTS) {
+    const date = new Date(event.period.to);
+    if (date.getTime() > now) {
+      events.push(event);
+    }
+  }
+  events.sort((a, b) => {
+    const t1 = new Date(a.period.from);
+    const t2 = new Date(b.period.to);
+    return t1.getTime() - t2.getTime();
+  });
+  return events;
 }
