+++ src/app/monthly-events/monthly-events.component.ts
@@ -1,11 +1,9 @@
-import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
-import { MatPaginator } from '@angular/material/paginator';
-import { MatSort } from '@angular/material/sort';
-import { MatTable } from '@angular/material/table';
+import { Component } from '@angular/core';
 
+import { MonthlyEventsDataSource } from './monthly-events-datasource';
+import { EventsComponent } from '../events.component';
 import { CommonService } from '../common.service';
-import { MonthlyEventsDataSource, EventInfo } from './monthly-events-datasource';
-import { BreakpointObserver, Observable, map, shareReplay, isHandset } from '../utilities';
+import { BreakpointObserver, Observable, map, shareReplay } from '../utilities';
 import { detailExpand } from '../animations';
 
 @Component({
@@ -14,14 +12,7 @@ import { detailExpand } from '../animations';
   styleUrls: ['./monthly-events.component.css'],
   animations: [detailExpand],
 })
-export class MonthlyEventsComponent implements AfterViewInit, OnInit {
-  @ViewChild(MatPaginator, {static: false}) paginator: MatPaginator;
-  @ViewChild(MatSort, {static: false}) sort: MatSort;
-  @ViewChild(MatTable, {static: false}) table: MatTable<EventInfo>;
-  dataSource: MonthlyEventsDataSource;
-  isHandset$: Observable<boolean>;
-  expandedElement: EventInfo | null;
-  showDetail = false;
+export class MonthlyEventsComponent extends EventsComponent {
 
   get displayedColumns$(): Observable<string[]> {
     return this.isHandset$.pipe(
@@ -31,34 +22,9 @@ export class MonthlyEventsComponent implements AfterViewInit, OnInit {
   }
 
   constructor(
-    private breakpointObserver: BreakpointObserver,
-    private cs: CommonService,
+    cs: CommonService,
+    breakpointObserver: BreakpointObserver,
   ) {
-    this.isHandset$ = isHandset(breakpointObserver);
-  }
-
-  ngOnInit() {
-    this.dataSource = new MonthlyEventsDataSource(this.cs);
-  }
-
-  ngAfterViewInit() {
-    this.dataSource.sort = this.sort;
-    this.dataSource.paginator = this.paginator;
-    this.table.dataSource = this.dataSource;
-  }
-
-  getDay(event: EventInfo): string {
-    return this.cs.getMonthlyDay(event.schedule);
-  }
-
-  getMonth(event: EventInfo): string {
-    return this.cs.getMonth(event.schedule);
-  }
-
-  getLocation(event: EventInfo): string {
-    const name = this.cs.getLocationName(event.location);
-    const location = this.cs.getLocation(event.location);
-    const region = this.cs.getPrefecture(location.prefecture);
-    return `${name}, ${region}`;
+    super(cs, new MonthlyEventsDataSource(), breakpointObserver);
   }
 }
