+++ src/app/upcoming-events/upcoming-events-datasource.ts
@@ -1,82 +1,33 @@
-import { DataSource } from '@angular/cdk/collections';
-import { MatPaginator } from '@angular/material/paginator';
-import { MatSort } from '@angular/material/sort';
-import { map } from 'rxjs/operators';
-import { Observable, of as observableOf, merge } from 'rxjs';
+import { AbstractEventsDataSource } from '../abstract-events-datasource';
+import { EventInfo } from '../models';
+import EVENTS from '../../assets/model/events.json';
 
-import { CommonService, EventInfo } from '../common.service';
 export { EventInfo };
 
 /**
- * Data source for the UpcomingEvents view. This class should
+ * Data source for the LocalEvents view. This class should
  * encapsulate all logic for fetching and manipulating the displayed data
  * (including sorting, pagination, and filtering).
  */
-export class UpcomingEventsDataSource extends DataSource<EventInfo> {
-  data: EventInfo[];
-  paginator: MatPaginator;
-  sort: MatSort;
-
-  constructor(private cs: CommonService) {
-    super();
-    this.data = this.cs.getUpcomingEvents();
-  }
-
-  /**
-   * Connect this data source to the table. The table will only update when
-   * the returned stream emits new items.
-   * @returns A stream of the items to be rendered.
-   */
-  connect(): Observable<EventInfo[]> {
-    // Combine everything that affects the rendered data into one update
-    // stream for the data-table to consume.
-    const dataMutations = [
-      observableOf(this.data),
-      this.paginator.page,
-      this.sort.sortChange
-    ];
-
-    return merge(...dataMutations).pipe(map(() => {
-      return this.getPagedData(this.getSortedData([...this.data]));
-    }));
-  }
-
-  /**
-   *  Called when the table is being destroyed. Use this function, to clean up
-   * any open connections or free any held resources that were set up during connect.
-   */
-  disconnect() {}
-
-  /**
-   * Paginate the data (client-side). If you're using server-side pagination,
-   * this would be replaced by requesting the appropriate data from the server.
-   */
-  private getPagedData(data: EventInfo[]) {
-    const startIndex = this.paginator.pageIndex * this.paginator.pageSize;
-    return data.splice(startIndex, this.paginator.pageSize);
-  }
-
-  /**
-   * Sort the data (client-side). If you're using server-side sorting,
-   * this would be replaced by requesting the appropriate data from the server.
-   */
-  private getSortedData(data: EventInfo[]) {
-    if (!this.sort.active || this.sort.direction === '') {
-      return data;
-    }
-
-    return data.sort((a, b) => {
-      const isAsc = this.sort.direction === 'asc';
-      switch (this.sort.active) {
-        case 'title': return compare(a.title, b.title, isAsc);
-        case 'date': return compare(+a.period.from, +b.period.from, isAsc);
-        default: return 0;
-      }
-    });
+export class UpcomingEventsDataSource extends AbstractEventsDataSource<EventInfo> {
+  constructor() {
+    super(eventData());
   }
 }
 
-/** Simple sort comparator for example ID/Name columns (for client-side sorting). */
-function compare(a, b, isAsc) {
-  return (a < b ? -1 : 1) * (isAsc ? 1 : -1);
+function eventData(): EventInfo[] {
+  const events: EventInfo[] = [];
+  const now = Date.now();
+  for (const event of EVENTS) {
+    const date = new Date(event.period.to);
+    if (date.getTime() > now) {
+      events.push(event);
+    }
+  }
+  events.sort((a, b) => {
+    const t1 = new Date(a.period.from);
+    const t2 = new Date(b.period.to);
+    return t1.getTime() - t2.getTime();
+  });
+  return events;
 }
